+++ eras/allegra/impl/CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### `testlib`
 
+* Added `EraSpecificSpec AllegraEra` instance
 * Added `Examples` module with: `ledgerExamples`, `exampleAllegraTxBody`, `exampleAllegraTxAuxData`
 * Added `Arbitrary` instance for `TransitionConfig AllegraEra`
 * Added `Era` module with `AllegraEraTest` class